Added more documentation, and refactored the implementations of AuthStore into separate packages.
